requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
requests
launchpadlib
python-dotenv